This is to update from master

+++ compiler/ballerina-lang/src/main/java/org/wso2/ballerinalang/programfile/DefaultValue.java
@@ -28,7 +28,7 @@ public class DefaultValue {
 
     public int valueCPIndex = -1;
     public long intValue;
-    public byte byteValue;
+    public long byteValue;
     public double floatValue;
     public String stringValue;
     public boolean booleanValue;
