+++ tests/ballerina-unit-test/src/test/java/org/ballerinalang/test/balo/types/FiniteTypeTest.java
@@ -18,9 +18,6 @@
 
 package org.ballerinalang.test.balo.types;
 
-import org.ballerinalang.launcher.util.BCompileUtil;
-import org.ballerinalang.launcher.util.BRunUtil;
-import org.ballerinalang.launcher.util.CompileResult;
 import org.ballerinalang.model.values.BBoolean;
 import org.ballerinalang.model.values.BByte;
 import org.ballerinalang.model.values.BFloat;
@@ -29,6 +26,9 @@ import org.ballerinalang.model.values.BMap;
 import org.ballerinalang.model.values.BString;
 import org.ballerinalang.model.values.BValue;
 import org.ballerinalang.test.balo.BaloCreator;
+import org.ballerinalang.test.util.BCompileUtil;
+import org.ballerinalang.test.util.BRunUtil;
+import org.ballerinalang.test.util.CompileResult;
 import org.testng.Assert;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
@@ -260,7 +260,7 @@ public class FiniteTypeTest {
         Assert.assertEquals(returns.length, 1);
         Assert.assertNotNull(returns[0]);
         Assert.assertSame(returns[0].getClass(), BByte.class);
-        Assert.assertEquals((((BByte) returns[0]).byteValue()), (byte) 222);
+        Assert.assertEquals((((BByte) returns[0]).byteValue()), 222);
     }
 
     @Test()
